Add autoscaling for Prometheus to prevent unschedule pods when the cluster is to small resulting in errors being reported.
